streaming http client, code cleanup
