+++ colossus/src/main/scala/colossus/protocols/http/HttpClient.scala
@@ -4,33 +4,39 @@ import scala.language.higherKinds
 
 import service._
 import scala.concurrent.{ExecutionContext, Future}
+import colossus.streaming.Source
 
 trait BaseHttpClient[M[_], B, P <: BaseHttp[B]] extends LiftedClient[P, M] {
+
+  def ops: MessageOps[HttpRequestHead, B, P#Request]// = implicitly[MessageOps[HttpRequestHead, B, P#Request]]
 
   protected lazy val hostHeader = clientConfig.map(config => {
     HttpHeader(HttpHeaders.Host, config.address.getHostName)
   })
+
+  override def send(input: P#Request): M[P#Response] = {
+    val headers = input.head.headers
+    val decoratedRequest = headers.firstValue(HttpHeaders.Host) match {
+      case Some(_) => input // Host header is already present.
+      case None => hostHeader.map(header => ops.withHeader(input,header)).getOrElse(input)
+    }
+    super.send(decoratedRequest)
+  }
+
 
 
 }
 
 trait HttpClient[M[_]]  extends LiftedClient[Http, M] with BaseHttpClient[M, HttpBody, Http] with HttpRequestBuilder[M[HttpResponse]]{ 
 
+  val ops = HttpRequestOps
 
   protected def build(req: HttpRequest) = send(req)
 
   val base = HttpRequest.base
 
-  override def send(input: HttpRequest): M[HttpResponse] = {
-    val headers = input.head.headers
-    val decoratedRequest = headers.firstValue(HttpHeaders.Host) match {
-      case Some(_) => input // Host header is already present.
-      case None => hostHeader.map(header => input.withHeader(header)).getOrElse(input)
-    }
-    super.send(decoratedRequest)
-  }
-
 }
+
 
 
 
@@ -45,12 +51,3 @@ object HttpClient {
   }
 
 }
-
-//trait StreamingHttpClient extends LiftedClient[StreamingHttp, Callback] with BaseClient[Callback, Source[Data]]
-
-/*
-object StreamingHttpClient {
-    
-  implicit object StreamingHttpClientLifter extends ClientLifter[
-}
-*/
